+++ Sources/BaseFoundationiOS/Validators/ValdiatorRules/SelectableRule.swift
@@ -1,0 +1,8 @@
+//
+//  File.swift
+//  
+//
+//  Created by pablo borquez on 05/02/24.
+//
+
+import Foundation